fix brightness for scale for aws and translate brightness of 0 to off cmd
fixes problems with AWS brightness scaling and add the capability to turn off the device if brightness is set to "0".  Currently if Govee device recieves a brightness of "0" on AWS or LAN or BLE it defaults to brightness of "1" and never turns off the device.  This code update fixes both the scaling issues reported by other users and turns off the devices when we turn the brightness down to "0".